Correct the objective's position.
Once the allied team catches the objective, if lost (only if he is killed), the objective coordinates
are set to those of the victim. If the objective is now caught by an axis member, it is set to its initial
position. The "flag_taken" belief is also removed from the allied soldier before he dies.